tests: Avoid 0-length pread confusion
The NBD protocol says 0-length reads are undefined.  We permit them
now, but are considering a knob to reject them client-side; but even
if we let it through to the server, we can't control whether the
server will reject our request for being 0-size.  When testing for
expected failures, it is better to use a valid buffer in combination
with whatever else we are intentionally botching, to ensure the
failure we detect is the one we expected rather than an unrelated
response to a 0-size request.

Fixes: adb11a32f3