python: Fix parsing of Bool parameters
We were passing &int to "b", but Python documents that for that
converter, PyArg_ParseTuple only sets the 'unsigned char' portion of
that address, which meant we often saw garbage in the remaining bytes
for a non-zero result even when the user passed False.  We could
either fix the type of the variable we are passing in, or we can
improve our code to use the "p" (predicate) parser instead, which was
added in python 3.3 (our minimum required version), which parses into
an 'int' and recognizes a wider array of python values.

(cherry picked from commit ea67475c5272db5841ea1f5391171cdac0f13e9e)
(cherry picked from commit 40060ed57d42bf9ddd15178e415a3cba6a51abfd)