ocaml: Fix Enum and Flags parameters
Two nasty coredumps, because we failed to test these two parameter types:

  NBD.set_tls nbd NBD.TLS.ALLOW;
triggers a call to TLS_val(v) with v set to NBD.TLS.t (an integer),
and crashes trying to deref it as an object (the code was copied from
Flags_val(), where v is 'NBD.TYPE.t list' and therefore each member of
the list does need a deref).

  NBD.set_handshake_flags nbd [];
failed to compile, because we typed it as 'NBD.HANDSHAKE_FLAG.t'
instead of 'NBD.HANDSHAKE_FLAG.t list'; but if you pass in a single
flag as typed, the code then calls HANDSHAKE_FLAG_val() which properly
tries to dereference that flag as a list.

Fix both code generation problems, as well as adding testsuite
coverage.  Note that we STILL have some bogus conversion to OCaml; the
ideal signature for get_* would be the same type that set_* took on
input, but right now we are blindly returning RInt (and have to
manually translate it).  Fixing that requires the addition of REnum
and RFlags to generator/API.ml.

(cherry picked from commit ac345ce3ec1ba4b54a232e7b576fb59ab115f326)