golang: Add test coverage of configuration knobs
Time to port the same tests that picked up bug fixes for python and
ocaml.  go feels so much more verbose, but at least the interfaces
seemed to work sanely without needing a generator fix.  It would be
nice if the get functions returned correctly typed constants instead
of bare uint, but that would require REnum and RFlags support in the
generator.