generator: Introduce REnum/RFlags return types
For now, there is no change to the generated code (although the man
pages are slightly improved), but this will make it easier for later
patches to improve OCaml and Golang bindings so that set/get pairs are
typed symmetrically.

This patch preserves a mismatch for RFlags: on input, we used
'uint32_t', but on output, where we previously used RUint, we produce
'unsigned'.  The two happen to be ABI compatible, so the next patch
will adjust that, at the expense of a minor API change on platforms
where 'uint32_t' is not the same rank as 'unsigned'.