generator: Refactor handling of closures in unlocked functions
We have a memory leak when a function with a closure exits early prior
to registering that closure through some path that will guarantee
cleanup.  The easiest way to fix it is to guarantee that once a
closure is passed into a public API, it will be cleaned regardless of
whether that API succeeds or fails.  But to avoid cleaning the closure
more than once, we need to refactor our internal handling, in order to
track when a closure has been handed off for later cleaning.  The
easiest way to do this is to pass closures by reference to all
internal functions, so that helpers in the call stack can modify the
incoming pointer rather than their own copy.

In turn, this means that we now have three different ways of including
a closure in a list of arguments: direct (such as in the public API
signature, with nothing added), as a pointer (declaring the internal
API, by adding '*'), or as a reference (building the call to the
internal API, by adding '&').

This patch is just preparatory, with no semantic change.  The public
API still passes closures by copy, but the generator then operates on
the address of that closure through all internal nbd_unlocked_*
functions, rather than making further copies through each additional
function call.

(cherry picked from commit ec6518d94ab8792e390999c501772abaf307976d)