api: s/unsigned/uint32_t/ for RFlags
On all platforms we care about, this is ABI-compatible.  However, it
is a potential API change for platforms where 'uint32_t' has different
rank than 'unsigned' (most likely on a 32-bit platform that uses
'long' for 'uint32_t').  That said, the fact that the generated
unlocked.h was using 'unsigned' for nbd_unlocked_get_handshake_flags,
while handle.c was using 'uint32_t', and we were not hitting any
compilation errors, means that no one was actually building on a
platform where this API difference was a problem.  Therefore, we
consider it safe to do.