python: Plug leak in test 600-debug-callback
As long as our debug callback remains registered, we have tied up a
memory allocation.  It's not strictly a leak, as it goes away with
nbd_close(), but it takes a tweak to get python to run the garbage
collector early enough to avoid seeing this allocation as a false
positive while I work on plugging actual leaks still present in nbdsh.