python: Plug some memory leaks on error paths
Inspection of the generated code showed several places where we did
not release references on all error paths.  In particular, switching
things to always jump to an out: label, even when the underlying C
function cannot fail, makes it easier to clean up when the user passes
wrong types to a function call.

One of the easiest triggers without this patch was this one-liner:
$ nbdsh -c 'h.block_status(512, 0, 1)'

which fails due to '1' not being callable, but leaked malloc'd memory
in the process.

(cherry picked from commit 82dac49af083eef023931c17413ea4339cbae6e4)