python: Fix more memory leaks
h.nbd_connect_command was leaking a python object per parameter, and
also leaks memory on failure.  In fact, it was possible to segfault on
something as trivial as:
nbdsh -c 'h.connect_command(["true",1])'

h.nbd_pread was leaking a read buffer on every single synchronous read.

My previous patch to address closure callbacks had a bug in
h.nbd_pread_structured and similar: if the allocation for Closure
fails, we hit 'goto err' prior to initializing the OClosure pointers,
which could lead to freeing garbage.  Similarly, when messing with
non-callable for either the Closure or the OClosure, there was enough
inaccuracy in python reference counting to cause a leak or
double-free.

We were not consistently checking for python function failures (such
as when a wrong type is passed in).

While touching this, drop a pointless cast to char* when calling
PyArg_ParseTuple, and rearrange some of the code for fewer loops over
args/optargs in the generator.

Fixes: 4fab2104ab
Fixes: bf0eee111f
Fixes: 82dac49af0