docs: Tweak wording on set_handshake_flag
Make it clear that since this API defaults to some bits set, and may
add more bits in the future, that it is best to use a
read-modify-write when disabling only one bit (as opposed to a blanket
set to 0 when disabling all bits).  Also fix a grammar issue.

(cherry picked from commit d95fe39043b39453c5ed97f18deeaf312fb5d137)