api: Add xxx_MASK constant for each Flags type
Document and make it easier for applications to determine which
bitmask values were supported at compile time.  Also, generating
bitmask values in hex tends to be more legible than in decimal, as it
makes it obvious that the values are intended for bitwise operations.
And it doesn't hurt that this reduces churn in some of the tests if
new bits are added.