docs: Fix stale comment in nbd_shutdown
When we were first designing the API, we had a flags argument for
nbd_aio_disconnect (always 0, per the current protocol spec), and no
flags for nbd_shutdown (which was designed to be a high-level wrapper
that does the right thing).  Then we changed our mind and decided we
may want to have a non-NBD-protocol flag argument for nbd_shutdown,
for example to allow a client-side failure if shutdown is not possible
if other commands are in flight, vs. forcing a shutdown in spite of
in-flight commands.  But this was all before we made the generator's
Flags type-specific, at which point we designated it as using
cmd_flags rather than an (as-yet empty) alternative bitmask.

We may _still_ decide that nbd_shutdown should use a different bitmask
than the cmd_flags passed to nbd_aio_disconnect (it won't change the C
API or ABI, but will have ripple effects to other language bindings
that are more strongly typed).  But until then, we should at least
remove the stale comment in our docs about not having a flags
parameter.

Fixes: 176fc4ea