api: Add LIBNBD_SHUTDOWN_ABANDON_PENDING flag
As mentioned in commits 176fc4ea and 609c25f0, our original plan in
adding a flags argument to nbd_shutdown was to let us specify
different behaviors at the libnbd level, rather than NBD protocol
flags (for that, the user has nbd_aio_disconnect).  But when we later
parameterized OFlags to accept various bitmasks (commit f891340b), we
failed to mark nbd_shutdown as using a different bitmask than
NBD_CMD_FLAG.

I've finally found a use for such a flag.  By itself,
nbd_aio_disconnect merely queues itself at the back of all pending
commands.  But if the server is processing responses slowly, it can be
desirable to elevate a disconnect request to the front of the queue,
intentionally abandoning all subsequent commands that have not already
started on their way to the server.

Fortunately, we have been rejecting all flag values, so changing the
type of the OFlags argument now has no impact to either the C API or
the ABI.  Other language bindings that are more strongly-typed will
see a different enum, but we don't promise compatibility there, and
even then, languages that permit omitting the flags argument in favor
of a default don't see any difference in clients that were omitting
the argument.  Still, it's wise to pick the value of our new flag to
be 0x10000 rather than 1, so that if in the future, we do decide to
let nbd_shutdown honor one or more NBD_CMD_FLAG_FOO, we can define
NBD_SHUTDOWN_FOO with the same values for less dynamic translation
between those future enum values.

Note that the new LIBNBD_SHUTDOWN_ABANDON_PENDING flag is not
necessarily instant: if the server is still receiving the previous
command, we have to wait for that before the server can learn of our
intent, and the command itself still blocks until we enter closed or
dead states.  But it can certainly reduce the time spent in
nbd_shutdown by not having to wait for all unsent commands in the
queue to also be processed by the server.

Other flags we might want to consider adding in the future:
- SHUTDOWN_IMMEDIATE: use shutdown(SHUT_WR) or similar to tell the
server we are not writing anything else, not even finishing an
NBD_CMD_WRITE that was stranded halfway through (although this is
effectively what you get with calling nbd_close()...)
- SHUTDOWN_WAIT_PENDING: prevent libnbd from queuing new data
commands, but don't actually send NBD_CMD_DISC to the server until
results to all pending commands have arrived, so that we don't risk
the server stranding an incomplete operation (although this is
effectively what you get by waiting for no pending commands yourself
before calling nbd_shutdown()...)