api: Add nbd_set_strict_mode with STRICT_COMMANDS
Right now, libnbd has refused to issue a command not advertised as
supported by a server, mainly because the NBD protocol does not
guarantee what the server will do, and libnbd would rather stay in
sync with the server than drop the connection.  However, for
integration purposes, it can be handy to coerce libnbd into sending
something to see how the server will react (whether it be an extension
libnbd has not yet learned, or an intentional bad request to test
server error handling).  Time to make this something the user can
control, by adding a new strictness mode.  Later patches will add
more knobs to the mode.