golang: Fix test 590
A golang closure of a shared variable sees the latest contents of that
variable, rather than the contents it had at the time the closure was
created.  We have to copy into another variable created with the same
lifetime as the closure, or else the test will try to write at the
wrong offsets.

More details on the problem and various workarounds:
https://www.ardanlabs.com/blog/2014/06/pitfalls-with-closures-in-go.html
http://devs.cloudimmunity.com/gotchas-and-common-mistakes-in-go-golang/index.html#closure_for_it_vars

I ran into the problem when adding offset validation - the shared soff
variable ultimately ends up at EOF after the last read_completed
callback, which resulted in every write after that point trying to
write beyond EOF and failing.