generator: Add default_flags
An upcoming patch wants to add an optional 'guard' member to some
flags constructs (for overriding whether the generator enforces that
only known flags are in the bitset at runtime).  Rather than having to
add this member to every flags, we can use a 'defaults with' construct
for the same amount of churn now, but with less future churn if we
come up with another member to add.

At the same time, older OCaml compilers issue:

Error (warning 23): all the fields are explicitly listed in this record:
the 'with' clause is useless.

when using a with clause but overriding all fields; the easiest
workaround is to have a dummy field that only the default sets.

Thanks: Rich Jones
(cherry picked from commit f03be9fcb144137975d9999e1cc93205d62f3700)