api: Add STRICT_FLAGS to set_strict_mode
The next strict knob: allow the user to pass unknown flags across the
wire (this is different than passing a known flag at the wrong time);
with this, we are back to client-side rejection of unknown flags as
the default.

It is interesting to note that NBD only permits 16 bits of flags, but
we have a signature that takes uint32_t; if we wanted, we could pack
libnbd-specific flags in the upper bits that the NBD protocol would
never see.