generator: Add optional guard clause when validating flags
We want to add another strictness flag controlling whether flags
unknown to libnbd are sent to the server.  But that in turn means we
need a way to bypass the generated code that rejects unknown flags in
the first place.  This patch adds an optional guard clause to all flag
types, and if it is present, an unknown bit is only rejected if the
guard condition is also met.  To test the code, the initial condition
is chosen to match the NBD protocol accepting only 16 bits of command
flags, where the code now rejects all bits larger than 16 but happens
to temporarily admit all smaller bits to the server instead of
performing client-side checking; it also requires that all commands
now pass flags all the way through the stack.  This will be tightened
back up to client-side rejection with a strictness knob in the next
patch.