nbdsh: Don't open-code logic for determining auto-debug
Rather than duplicating an env-var check ourselves (which gets out of
date if libnbd ever changes how it determines when to do debugging by
default), it is easier to just create a new NBD handle and check if it
has debugging enabled.  Note that we can't rely on the 'h' that we
just opened prior to executing -c statements, as the user may have
closed it, but fresh handles are cheap.

Thanks: Rich Jones