maint: Consistent rules for nbdcopy/info when libxml is absent
Any application that requires URI support is not worth installing when
libnbd is built with './configure --disable-libxml2', because the app
would only ever fail.  We can still build the applications locally to
detect bit-rot, but that's all the further we need to go.  And if the
app is not installed, we don't need the man pages or bash completion
hooks.