api: Add get_nr_meta_contexts, clear_meta_contexts
An upcoming commit will add the ability to query which meta contexts a
server is advertising; but first we need to be able to present a
different list of meta context requests in response to what we learn
from the server.  Add APIs for reviewing which requests have already
been registered, as well as a way to reset the list of requests.

This adds some testsuite coverage; and more will appear with the
addition of nbd_opt_list_meta_contexts in later patches.