info/info-json.sh: Make the test robust.
The test used nbdkit-floppy-plugin to build a disk image from the
current directory, and it then ran a trivial test against it (not
using anything in the disk image).

This was not robust when the test was run in parallel because the
current directory could change rapidly, resulting in very rare test
failures if a temporary file was created, seen by the floppy plugin,
and then destroyed when the plugin tried to read it.

Fix this by using nbdkit-memory-plugin.

(cherry picked from commit 23f22618f01ce9c170bf24656501457094d3dc86)