nbdsh: Add --opt-mode command line option
Similar to --base-allocation, this is a common enough configuration
that cannot be performed with -c when using --uri, but which makes
life easier in scripting.  And like --base-allocation, I'm starting
with only a long option spelling, rather than burning a short option
letter.

In fact, with --opt-mode, we no longer have the urgency to keep
--base-allocation, because you can set that during option negotiation
(that is, setting opt mode before connection enables you defer all
other steps that used to require pre-connection actions to instead be
during negotiation); but there's no harm in keeping --base-allocation
around for back-compatibility.