tests: Adjust opt-list for latest nbdkit
As previously noted in our tests, nbdkit .list_exports is not a stable
interface until nbdkit 1.24.  As such, recent nbdkit commit 3c06520edc
decided that in the name of security, allowing more than one
NBD_OPT_LIST per connection is too risky for abuse as an amplification
attack (for a lengthy list, the server sends far more reply traffic
than the client requested, which can be abused if the client is not
serious about connecting), and this breaks our tests that repeat it on
a single connection.

Hence, we have to fix opt-list (and its variants in other languages)
to use means other than repeated nbd_opt_list() calls on a single
connection to test the various control paths.

For python, the changes to the test are more concise by use of a
context manager.  For ocaml and go, I opted for a helper function that
takes an anonymous function argument.  Hopefully my lack of
familiarity with language idioms does not show too badly through my
efforts to avoid pointless repetition of the setup/teardown code.

(cherry picked from commit 6f73b530f7cc5280bb7afd4dc5a6913c7f3573a4)