info: Write output atomically.
If the server fails, nbdinfo can write partial output before the error
message (albeit on different channels).  Here is an example:

  $ nbdkit eval open='echo EIO fail >&2; exit 1' --run 'nbdinfo --json "$uri"'
  {
  "protocol": "newstyle-fixed",
  "TLS": false,
  nbdkit: eval[1]: error: /tmp/nbdkitii3pZW/open: fail
  nbdkit: eval[1]: error: /tmp/nbdkitii3pZW/open: fail
  nbdinfo: nbd_opt_go: server replied with error to opt_go request: No such file or directory

Notice the partial JSON document, and the error message from nbdinfo.

With this commit nbdinfo tries to produce either the complete output
or the error message.

Thanks: Eric Blake