interop: Let dirty-bitmap.sh run in VM
When testing libnbd in an L1 VM where nested virt was not set up,
qemu-kvm failed because /dev/kvm was missing and unable to create an
L2 VM.  But we aren't running a guest; merely using the bare minimum
to set up qemu-kvm as an NBD server.  So allowing a scaled down
machine makes the test pass.

Thanks: Daniel P. Berrangé