info: Adjust to actual 'qemu-nbd -A' semantics
Review on the qemu list has led to an altered definition of what
'qemu:allocation-depth' should report: rather than a tri-state value,
it is an actual depth.  It's time to match what actually got committed
into qemu, which in turn means a slight refactoring to use a malloc'd
string for a description.  And testsuite coverage never hurts.

Fixes: 71455c021