info: --map: Use nbdkit vector of uint32_t to implement extents.
This refactoring simply changes the code so that the extent callback
builds up a list of extents (using a vector of uint32_t) and then
there is a separate function to print, instead of the callback doing
the printing.  This will make coalescing of entries -- next commit --
easier to implement.