copy: Allowing copying from NBD server to NBD server.
This is a general refactoring of nbdcopy in preparation for future
work, incidentally adding the feature of copying between NBD servers
which was missing before.

We abstract the source and destination into structures that can
describe either a local (file/device/pipe/stdio) or an NBD server.
The copy routine can copy between generic source and destination
instead of having separate upload and download functions.