copy: Add --synchronous flag.
This flag forces synchronous copying using the high level API.  This
mode is also forced when streaming to and from stdio, pipes and
sockets.

Since this is the only type of copying that nbdcopy can do at the
moment, this flag currently does nothing.  It will make more sense
when we add threads.