copy: Implement machine-readable progress bar.
This also makes the progress bar code thread-safe which it wasn't
really before (I don't think it was unsafe but you might have got
mixed up output).

This also makes it print 100/100 once at the end.  In the multi-
threaded case it would do that once per thread which was wrong.