python: Fix failures in test python/t/590-aio-copy.py
This test failed (rarely) with an assertion failure indicating that
the full data had not been copied.  Note this failure usually only
happens on slow machines (eg. Koji i686), and is very hard to
reproduce on x86-64.

This seems to be because the main loop could exit early -- in
particular if there were either pending commands in flight on the
source handle, or pending write commands in the "writes" list.  By
tightening up the condition on the main loop we should avoid these
failures.