copy: Correct file_get_extents.
The get_extents function is supposed to return precisely bounded
extents, adjusting between the underlying file and NBD APIs and what
the main loop expects.  We did this for NBD but not for file.

The result of this was over-copying of data (although not corruption)
when copying from a file.  In the case of files with holes larger than
4G you could also see an assertion error because we tried to issue
a > 4G trim.