copy: Share slices instead of copying when splitting commands.
When we do sparseness detection we may end up splitting a write
command into several (ie. data and holes).  When I profiled this in
perf I consistently saw large amounts of time (25%+) spent memcpy-ing
data from the command to the new sub-commands.

Instead of copying buffers, introduce a notion of slices which provide
a shared view of the underlying data buffer.

This also changes some nomenclature (commands were previously called
"buffers" which was not very helpful).