copy/file-ops.c: Port zero strategy from nbdkit
Port the zero strategy from nbdkit file plugin, improving the
reliability and compatibility with block devices on modern kernels.

struct file_rw has now capability flags: can_punch_hole, can_zero_range,
can_fallocate, and can_zeroout. The flags are initialized based on the
type of the file descriptor and compile time checks:

- For regular file, we enable can_punch_hole, can_zero_range, and
  can_fallocate.
- For block device, we enable can_punch_hole, can_zero_range, an
  can_zeroout.

When calling zero() in the first time, we try the following methods,
returning on the first success:

- If don't need to allocate, try to punch a hole.
- Try to zero the range
- Try to combine punching a hole and fallocate
- Try BLKZEROOUT ioctl

If a method is not supported, we disable the capability flag, so the
next call can try only what works.

The fallocate and ioctl wrappers return false when the call is not
supported by the underlying storage so we can disable the capability.
Previously the process would exit with an error.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>