copy: Preserve the host page cache when reading from local files.
When reading from a local file we can take advantage of the page cache
(ie. not having to read the file from disk if a copy is present in
memory), while at the same time not disturbing the state of the page
cache.  Disturbing the page cache can have bad consequences for other
processes running on the host since they will have their working set
evicted so it is something we should generally avoid.

This requires Linux APIs, using the technique described here:
https://insights.oetiker.ch/linux/fadvise/

This change only affects reads, since doing the same for writes is
even more complicated.

You can see the effect using the tools from
https://github.com/Feh/nocache

Before this change:

  $ cachestats /var/tmp/random
  pages in cache: 315648/8388608 (3.8%)  [filesize=33554432.0K, pagesize=4K]
  $ ./run time nbdcopy /var/tmp/random null:
  1.80user 23.70system 0:20.69elapsed 123%CPU (0avgtext+0avgdata 135084maxresident)k
  64585232inputs+0outputs (6major+8391948minor)pagefaults 0swaps
  $ cachestats /var/tmp/random
  pages in cache: 7734356/8388608 (92.2%)  [filesize=33554432.0K, pagesize=4K]

Notice that a large part of the file has been loaded into the page
cache after the run.

After this change:

  $ cachestats /var/tmp/random
  pages in cache: 315008/8388608 (3.8%)  [filesize=33554432.0K, pagesize=4K]
  $ time ./run nbdcopy /var/tmp/random null:
  real	0m21.602s
  user	0m1.944s
  sys	0m29.013s
  $ cachestats /var/tmp/random
  pages in cache: 315648/8388608 (3.8%)  [filesize=33554432.0K, pagesize=4K]

Notice there is only a small increase in the amount of file which is
cached, the elapsed time is about the same, but there is an increase
in system time (presumably the overhead of POSIX_FADV_DONTNEED).