copy: Evict pages from the page cache when writing to local files.
When writing to a file or block device, we are always writing new
(ie. previously uncached) data.  This commit ensures that very little
of that data will be in the page cache after nbdcopy finishes by
evicting it as we go along.  This ensures that the host page cache is
largely unchanged for other host processes.

This uses Linus's technique described here:
https://stackoverflow.com/a/3756466
but instead of using 2 windows, it uses a configurable larger number
of windows (in this case 8).

Before this commit:

  $ rm /var/tmp/pattern ; sync ; time ./run nbdcopy [ nbdkit pattern 32G ] /var/tmp/pattern && cachestats /var/tmp/pattern
  real	0m34.852s
  user	0m18.368s
  sys	0m33.117s
  pages in cache: 7090389/8388608 (84.5%)  [filesize=33554432.0K, pagesize=4K]

Notice that the newly written file ends up in the cache, thus trashing
the page cache on the host.

After this commit:

  $ rm /var/tmp/pattern ; sync ; time ./run nbdcopy [ nbdkit pattern 32G ] /var/tmp/pattern && cachestats /var/tmp/pattern
  real	0m38.721s
  user	0m18.837s
  sys	0m40.654s
  pages in cache: 65536/8388608 (0.8%)  [filesize=33554432.0K, pagesize=4K]

The newly written file does not disturb the page cache.  However there
is about 11% slow down.