copy: Simplify extents handling.
We don't have to worry about the server returning an extent either
partially or completely before the requested offset - this is
impossible because of the format of the NBD_REPLY_TYPE_BLOCK_STATUS
message.  We can therefore delete some dead code.  I replaced it with
a simple assertion in case there's something we have overlooked.

This also tightens the condition that the server is making progress -
previously a server could have returned a non-empty list of
zero-length extents causing nbdcopy to spin.

Thanks: Eric Blake
Fixes: commit 7fb099adfbc9ee6626c3aa92254b9a91cfac12e9