Revert "copy: file-ops.c: Remove unneeded check"
This reverts commit 0f6e4f38bc440fc52c20a3a448ef031f806ec5e2.

This fails on ppc64le only with:

lt-nbdcopy: file-ops.c:136: page_was_cached: Assertion `page < rwf->cached_pages.size' failed.

Coincidentally this is the only architecture on Fedora that uses a 64K
page size, although I don't know if that is connected.  I was not able
to make this fail on x86.