copy/file-ops.c: Fix page eviction when len < page_size.
On Fedora ppc64le at the moment page size is 64K.  When asked to evict
a range with length < 64K the length calculation wrapped around and it
tried to evict a huge number of pages beyond the end of the file.

With Nir's commit 0f6e4f38b this (correctly) resulted in an assertion
failure.

Fix this by checking for the overflow and returning early.