examples: copy-libev.c: Improve polling
Based on lib/poll.c, we should not invoke both notify_read() and
notify_write() on the same event, since the first call may change the
handle state and invalided the write event.

When having both read and write event, prefer the read event since it is
is a reply for older command.

When returning events, use switch case for more efficient conversion
from libnbd events to libev events.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>