tests: Set minimum compiler to ISO C99.
Previously we at least claimed that libnbd could be used (not
compiled) with an ANSI (ie. C89/C90) compiler.  However this is not in
fact true because we require the C99 boolean type.

Therefore raise the minimum compiler to ISO C99.

Revealed by compiling with clang and -std=c90 which is more picky than GCC.