interop: Tolerate qemu-nbd 6.0 change in behavior
Qemu 6.0 commit 0da9856851 (nbd: server: Report holes for faw images)
intentionally changed how holes are reported, not only for raw images,
but for known-zero portions of qcow2 files.  As the point of our
dirty-bitmap test is more about reading multiple contexts from a
single qemu-nbd process, it is okay to accept both behaviors for
base:allocation reporting on a block of zeros.