fuse: tests: Check fusermount3 is present before running tests
On Fedora fusermount3 is not present in the fuse3-libs package, so it
is possible to compile nbdfuse but not have fusermount3 installed.
The tests will fail in this case because they are unable to unmount
the temporary fuse mountpoints that are created.

(Actually the same problem was present even for fuse2 / fusermount,
but we didn't hit it before because most people would have fuse +
fuse-libs installed at the same time.)

Fixes: commit c74c7d7f01975e708b510e518895088fc61b5623