Introduction of automated testing
This setup uses the libvirt-ci to get as much coverage as possible with the ease
of use provided by that repository.  The changes required in libvirt-ci are
included in this merge request:

  https://gitlab.com/libvirt/libvirt-ci/-/merge_requests/120

Details about how the CI works and how can the builds be reproduced locally is
described in ci/README.rst.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>