fuse: Enable multithread support
This was much simpler than I though, merely changing fuse_loop ->
fuse_loop_mt.  FUSE itself will call our operations in parallel, so we
have to be careful about shared state.

Testing with fio shows this is about the same or a tiny bit slower
than before, which is not that surprising -- we are using the
synchronous API.  It should be possible to improve performance by
issuing commands in parallel on the same handle [see next commit]
and/or by using multi-conn.

test.fio contains:

  [global]
  filename=test.img
  bs=4k
  direct=1
  rw=read

  [job1]

  [job2]

  [job3]

  [job4]

Run the test:

  $ touch test.img
  $ nbdfuse test.img [ nbdkit memory 1G ] &
  $ fio test.fio
  $ fusermount3 -u test.img

Without this patch:

   READ: bw=239MiB/s (250MB/s), 59.6MiB/s-59.7MiB/s (62.5MB/s-62.6MB/s), io=4096MiB (4295MB), run=17166-17167msec

With this patch:

   READ: bw=237MiB/s (248MB/s), 59.2MiB/s-59.2MiB/s (62.0MB/s-62.1MB/s), io=4096MiB (4295MB), run=17303-17310msec