fuse: Issue commands in parallel
On top of the previous commit which enabled multithreading but
continued to use the synchronous libnbd API, this allows each thread
to issue commands asynchronously.  Because there is still a single
handle, this introduces a single background thread to poll the file
descriptor and dispatch the commands.

This is only a little bit faster (compare to results in previous
commit message):

   READ: bw=250MiB/s (262MB/s), 62.4MiB/s-62.4MiB/s (65.4MB/s-65.5MB/s), io=4096MiB (4295MB), run=16398-16411msec

A future multi-conn version of nbdfuse would likely use multiple
background threads (one per connection) to do the same job, but that
is left for future work.